+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [6.4.0-preview] - 2019-XX-XX
+### Added
+- Addressing mode for Sequential blocks
 
 ## [6.3.0-preview] - 2019-02-18